Updated mysql cookbook and added mass db/user pair creator

+++ cookbooks/mysql/recipes/default.rb
@@ -1,28 +1,32 @@
-#data_bag("rtb_data_bag")
-#db = data_bag_item("rtb_data_bag", "rtb")
-#password = db[node.environment]['mysql']['password'] || "qqqqqq"
-password = "qqqqqq" # default password, need to set up override
+mysql_initial = data_bag_item("mysql", "initial")
+password = mysql_initial["root_password"]
 
-bash "install_mysql" do
-  user "root"
-  cwd "#{Chef::Config[:file_cache_path]}"
+bash "mysql_initial_install" do
   code <<-EOH
     export DEBIAN_FRONTEND=noninteractive
     apt-get -q -y install mysql-server
   EOH
-  action :run
-  #not_if {File.exists?("#{Chef::Config[:file_cache_path]}/mysql_lock")}
+  user "root"
+  not_if "test -f #{Chef::Config[:file_cache_path]}/mysql_lock"
 end
 
-bash "config_mysql" do
-  user "root"
-  cwd "#{Chef::Config[:file_cache_path]}"
+bash "mysql_initial_config" do
   code <<-EOH
     mysqladmin -u root password #{password}
     echo "grant all privileges on *.* to 'root'@'%' identified by '#{password}'; FLUSH PRIVILEGES;" | mysql -u root -p#{password}
   EOH
-  action :run
-  #not_if {File.exists?("#{Chef::Config[:file_cache_path]}/mysql_lock")}
+  user "root"
+  not_if "test -f #{Chef::Config[:file_cache_path]}/mysql_lock"
 end 
 
-#mysql lock file
+package "mysql-server" do
+  action :upgrade
+  only_if "test -f #{Chef::Config[:file_cache_path]}/mysql_lock"
+end
+
+file "#{Chef::Config[:file_cache_path]}/mysql_lock" do
+  owner "root"
+  group "root"
+  mode "0755"
+  action :touch
+end
